requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Login==0.6.2
python-dotenv==1.0.0
bcrypt==4.0.1
python-docx==0.8.11
matplotlib==3.7.2
pandas==2.0.3
pymongo==4.5.0
Werkzeug==2.3.7
numpy<2
